Created these scripts from the tutorial
They are not sufficient for us to complete the assignment, but at least give us a base to start from, which we can build from. It is not yet commented, and it may not be 100% good. it seems to work, but movement seems funky, so I made a mistake somewhere.